better style and consts
better style and consts

+++ TCPListener/TCPListener/Program.cs
@@ -14,11 +14,13 @@ namespace TCPListener
 
 
             // Initialisation
-            byte[] clientIp = new byte[256];
+            byte[] clientIpBytes = new byte[256];
             int numberIpBytes = 0;
             int port = 13000;
-            string ip = string.Empty;
-            IPAddress localIp = IPAddress.Parse("192.168.1.172");
+            string clientIp = string.Empty;
+            const string serverIp = "10.5.48.44";
+
+            IPAddress localIp = IPAddress.Parse(serverIp);
 
             // Build the server
             server = new TcpListener(localIp, port);
@@ -26,18 +28,21 @@ namespace TCPListener
 
             TcpClient client = null;
             NetworkStream stream = null;
+            Console.WriteLine("Server opened ! \n");
 
             // Do it for each client that connected
             while (true)
             {
-                getClientIp(ref client, ref stream, ref server, ref numberIpBytes, ref clientIp, ref ip);
+                getClientIp(ref client, ref stream, ref server, ref numberIpBytes, ref clientIpBytes, ref clientIp);
             }
         }
 
-        static void getClientIp(ref TcpClient client, ref NetworkStream stream, ref TcpListener server, ref int numberIpBytes, ref byte[] clientIp, ref string ip)
+        static void getClientIp(ref TcpClient client, ref NetworkStream stream, ref TcpListener server, ref int numberIpBytes, ref byte[] clientIpBytes, ref string clientIp)
         {
             try
             {
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.WriteLine("Waiting for a client... \n");
                 client = server.AcceptTcpClient();
                 stream = client.GetStream();
                 Console.WriteLine("New client found ! \n");
@@ -48,31 +53,36 @@ namespace TCPListener
 
 
                 stream.Write(bytes, 0, bytes.Length);
-                Console.WriteLine("Message sent !");
+                Console.ForegroundColor = ConsoleColor.Blue;
+                Console.WriteLine("Message sent: " + command);
 
                 // Get IP and create file but only if an IP has been returned
-                numberIpBytes = stream.Read(clientIp, 0, clientIp.Length);
+                numberIpBytes = stream.Read(clientIpBytes, 0, clientIpBytes.Length);
 
                 if (numberIpBytes != 0)
                 {
-                    ip = System.Text.Encoding.ASCII.GetString(clientIp);
+                    clientIp = System.Text.Encoding.ASCII.GetString(clientIpBytes);
 
-                    Console.WriteLine("Client IP: " + ip);
+                    Console.WriteLine("Client IP: " + clientIp);
 
                     string fileName = "ip.txt";
 
-                    File.Delete(fileName);
-                    File.Create(fileName).Close();
-                    File.AppendAllText(fileName, ip + "\n");
+                    if (!File.Exists(fileName))
+                    {
+                        File.Create(fileName).Close();
+                    }
+                    File.AppendAllText(fileName, clientIp + "\n");
                 }
                 else
                 {
+                Console.ForegroundColor = ConsoleColor.Red;
                     Console.WriteLine("No IP was found");
                 }
             }
             catch (Exception ex)
             {
-                Console.WriteLine("Error: " + ex.Message);
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine("\nError: " + ex.Message + "\n");
             }
         }
     }
